requirements.txt
google-adk
google-genai
chainlit>=1.1.0
pandas
presidio-analyzer>=2.2.35
presidio-anonymizer>=2.2.35
spacy>=3.7.4
pytest
pytest-asyncio
python-dotenv>=1.0.0